requirements.txt
dlib==19.15.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.0.2
gunicorn==19.9.0
numpy==1.14.3
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
Pillow==5.1.0
